junior_chapter_2/src/main/java/ru/job4j/io/oracle/Server.java
package ru.job4j.io.oracle;

import java.io.BufferedReader;
import java.io.IOException;
import java.io.InputStreamReader;
import java.io.PrintWriter;
import java.net.ServerSocket;
import java.net.Socket;

public class Server {

    Socket socket;

    public Server(Socket socket) {
        this.socket = socket;
    }

    public void start() throws IOException {
        PrintWriter out = new PrintWriter(socket.getOutputStream(), true);
        BufferedReader in = new BufferedReader(new InputStreamReader(socket.getInputStream()));
        String ask;
        do {
            System.out.println("wait command ...");
            ask = in.readLine();
            System.out.println(ask);
            if ("Hello oracle".equals(ask)) {
                out.println("Hello, dear friend, I'm a oracle.");
                out.println();
            } else if (!"exit".equals(ask)) {
                out.println("I dont know.");
                out.println();
            }
        } while (!"exit".equals(ask));
    }

    public static void main(String[] args) {
        try (Socket socket =  new ServerSocket(5000).accept()) {
            new Server(socket).start();
        } catch (Exception e) {
            e.printStackTrace();
        }
    }
}
